Split review_pipeline.py god-module (3104 LOC) into 7 cohesive siblings
Move each review pipeline section into its own module under python/larch/review/:
- review_pipeline_shared.py: shared utilities and data types
- review_gather.py: gather-context
- review_prune.py: reviewer-prune
- review_dispatch_panel.py: dispatch-panel and dynamic archetypes
- review_collect.py: collect-findings
- review_threshold.py: check-reviewer-failure-threshold
- review_core_body.py: review-core pipeline orchestration

review_pipeline.py becomes a thin backward-compat re-export shim.
Update complexity-baseline.json and ruff.toml, update test imports for
monkeypatching in the correct sub-module namespace.

No behavior change; make py-lint green, 310 tests pass.
